# app/Models/User.php
<?php

namespace App\Models;

// use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;
use Laravel\Fortify\TwoFactorAuthenticatable;
use Laravel\Jetstream\HasProfilePhoto;
use Laravel\Sanctum\HasApiTokens;

/**
 * @OA\Schema(
 *     schema="User",
 *     title="User",
 *     description="User model",
 *     required={"name", "email", "password"},
 *     @OA\Property(
 *         property="name",
 *         type="string",
 *         description="The name of the user"
 *     ),
 *     @OA\Property(
 *         property="email",
 *         type="string",
 *         format="email",
 *         description="The email address of the user"
 *     ),
 *     @OA\Property(
 *         property="password",
 *         type="string",
 *         format="password",
 *         description="The password of the user"
 *     )
 * )
 */


class User extends Authenticatable
{
    use HasApiTokens;
    use HasFactory;
    use HasProfilePhoto;
    use Notifiable;
    use TwoFactorAuthenticatable;

    /** 
     
*The attributes that are mass assignable.
*
*@var array<int, string>
*/
protected $fillable = [
    'name',
    'email',
    'password',];

   /**
     
*The attributes that should be hidden for serialization.
*
*@var array<int, string>
*/
protected $hidden = [
    'password',
    'remember_token',
    'two_factor_recovery_codes',
    'two_factor_secret',
];

    /** 
     
*The accessors to append to the model's array form.
*@var array<int, string>
*/

protected $appends = [
    'profile_photo_url',
];

     /**
     
*Get the attributes that should be cast.
*@return array<string, string>
*/


protected function casts(): array{
    return ['email_verified_at' => 'datetime','password' => 'hashed',];}
}